Add group clone command
Use case: I have a set of "default" groups that contain commonly used repos.
But work on a new project demands I add some repos to a group.
Instead of modifying the default groups, create a new group based on the default
(clone it) and add new repos to the clone.