Return failure code to the system when commands fail
Currently mu will always return 0 to the system even when the git commands have failed.

While it is definitely good to always try the operations in all repositories (instead of stopping at the first error), always returning 0 to the system makes using 'mu' in scripts difficult, as we cannot easily be sure that the operations were successful.

In addition, this will now print (in red) the repositories where the operation has failed: a very common beginner mistake when using `mu` is to not notice that the command failed for one of the repositories, as it may end up at the top of the output.